+++ processor/src/main/java/galdr/processor/Generator.java
@@ -19,6 +19,7 @@ import javax.annotation.processing.ProcessingEnvironment;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.Modifier;
 import javax.lang.model.element.TypeElement;
+import org.realityforge.proton.GeneratorUtil;
 
 final class Generator
 {