+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Include the location of the problem when generating warnings for unnecessary public or protected access on members.
+* Use the `org.realityforge.proton:proton-processor-pack` project for processor utility methods.
 
 ### [v0.02](https://github.com/realityforge/galdr/tree/v0.02) (2019-11-30)
 [Full Changelog](https://github.com/realityforge/galdr/compare/v0.01...v0.02)